% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_process_plot.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a character value}
\usage{
capitalize(x)
}
\arguments{
\item{x}{A \code{character} string to be capitalized.}
}
\value{
The input string, but where the first letter is uppercase.
}
\description{
Capitalize a character value
}
\examples{
capitalize("norway")
}
