% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_data.R
\name{expand_data}
\alias{expand_data}
\title{Expand Data to Include Entries Inbetween Missing Dates}
\usage{
expand_data(data, id_cols = NULL, date_col = NULL, measurement_cols = NULL)
}
\arguments{
\item{data}{The dataset to be expanded. Must be a \code{data.table} object. Are attempted converted if it is not a \code{data.table} object.}

\item{id_cols}{A \code{character} vector of column names to group by. All column names included must be part of \code{data}.}

\item{date_col}{A \code{character} value corresponding to the name of the date column. Must be part of \code{data}.}

\item{measurement_cols}{A \code{character} vector corresponding to the names of the measurement columns. All column names included must be part of \code{data}. Currently not used.}
}
\value{
A \code{data.table} object that is the expanded version of the input \code{data}
}
\description{
Expand Data to Include Entries Inbetween Missing Dates
}
\examples{
print(1)
}
