% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_process_plot.R
\name{get_date_tick_distance}
\alias{get_date_tick_distance}
\title{Get Date Breaks Using The Range of Dates}
\usage{
get_date_tick_distance(date_range)
}
\arguments{
\item{date_range}{An \code{integer}.}
}
\value{
The date breaks
}
\description{
Get Date Breaks Using The Range of Dates
}
\examples{
get_date_tick_distance(39)
}
