% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_process_plot.R
\name{get_default_plotting_arguments}
\alias{get_default_plotting_arguments}
\title{Convert From Default to Their Corresponding Values}
\usage{
get_default_plotting_arguments(
  plotting_arguments,
  date_range,
  measure = c("median", "hyper", "hypo"),
  warning = c("slope", "bias", "peer_group")
)
}
\arguments{
\item{plotting_arguments}{A named \code{list} of plotting arguments}

\item{date_range}{An \code{integer}.}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{warning}{A \code{character} signifying which warning process that is of interest to simulate.}
}
\value{
A named \code{list} with the actual plotting arguments.
}
\description{
Convert From Default to Their Corresponding Values
}
\examples{
get_default_plotting_arguments(list("warning_point_shape" = 3))
}
