% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_gaps}
\alias{get_gaps}
\title{Get Gaps for Smoothing Window}
\arguments{
\item{x}{A \code{numeric} vector containing either medians, hyper percentages or hypo percentages used for the smoothing window.}

\item{get_grouping}{A non-missing \code{logical} value. If \code{TRUE}, the function returns an \code{integer} vector that describes the grouping of the x-values based on the gaps observed. Overrides \code{get_last_group}.}

\item{get_last_group}{A non-missing \code{logical} value. If \code{TRUE}, the function returns an \code{integer} vector containing the indicies of the last group based on the gaps observed in the x-values. \code{get_grouping} must be set to \code{FALSE} in order for this to work.}
}
\value{
A \code{integer} vector containing relevant gap information based on \code{get_grouping} and \code{get_last_group} arguments.
}
\description{
Efficiently provide information about the gaps in the measurements used for kernel smoothing. Used to validate whether we have enough information to give a slope warning to the user.
}
\details{
Get Gaps for Smoothing Window

Cannot locate gaps in x if dates with missing values are stripped. If both \code{get_grouping} and \code{get_last_group} is \code{FALSE}, only the gap counts are returned.
}
\examples{
\dontrun{

  print(1)

}
}
