% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_smoothing.R
\name{kernel_smoothing}
\alias{kernel_smoothing}
\title{Kernel Smoothing for Multiple Instruments Within Multiple Laboratories}
\usage{
kernel_smoothing(
  data,
  by,
  method = "lc",
  measure = c("median", "hyper", "hypo"),
  bw = 11,
  average = 0,
  standard_deviation = 1,
  attach = TRUE,
  approximate = TRUE,
  matrix_approach = TRUE,
  na_rm = TRUE
)
}
\arguments{
\item{data}{A \code{data.table} containing measurements.}

\item{by}{A \code{character} vector containing the names of the grouping variables.}

\item{method}{A \code{character} signifying which smoothing method that is used. Valid options include \code{lc} (Local-Average kernel smoothing) and \code{ll} (Local-Linear kernel smoothing)}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{bw}{A \code{double}. The bandwidth used for the smoothing.}

\item{average}{A \code{double}. For calculation of the smoothed median curve slopes. Should be equal to 0 for hypo and hyper.}

\item{standard_deviation}{A \code{double}. For calculation of the smoothed median curve slopes. Should be equal to 1 for hypo and hyper.}

\item{attach}{A \code{logical} value. Should we attach the information found in \code{data} to the output?}

\item{approximate}{A \code{logical} value. If set to \code{TRUE}, second differences are used to estimate gradients. If \code{FALSE} other methods are utilized.}

\item{matrix_approach}{A \code{logical} value. Should matrix calculus be used to smooth. Only relevant if \code{method} is \code{ll}. Matrix calculus is faster, but may not be stable in some cases.}

\item{na_rm}{A \code{logical} value. Should rows with missing measurements be stripped from the output?}
}
\value{
A \code{data.table} containing the smoothed values for multiple instruments within multiple laboratories
}
\description{
Kernel Smoothing for Multiple Instruments Within Multiple Laboratories
}
\examples{
print(1)
}
