% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_smoothing_check.R
\name{kernel_smoothing_check}
\alias{kernel_smoothing_check}
\title{Check if we have enough observations to do kernel smoothing correctly}
\usage{
kernel_smoothing_check(
  data,
  by,
  measure = c("median", "hyper", "hypo"),
  tol = 2,
  only_acceptable = TRUE
)
}
\arguments{
\item{data}{A \code{data.table} containing measurements.}

\item{by}{A \code{character} vector containing the names of the grouping variables.}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{tol}{A \code{integer}. The minimum acceptable number of results. Only groups with more than this value will be acceptable if \code{only_acceptable} is set to \code{TRUE}.}

\item{only_acceptable}{A \code{logical} value. If set to \code{TRUE}, only the grouping variables with enough observations are presented. Otherwise \code{FALSE} is returned by the function. If set to \code{FALSE}, a \code{data.table} object containing relevant information is outputted.}
}
\value{
A \code{logical} value or a \code{data.table} object depending on \code{data}.
}
\description{
Check if we have enough observations to do kernel smoothing correctly
}
\examples{
print(1)
}
