% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{laks_loo}
\alias{laks_loo}
\title{Local Average Kernel Smoothing Leave One Out Cross Validation}
\arguments{
\item{date}{A \code{numeric} vector containing dates for the smoothing.}

\item{median}{A \code{numeric} vector contaning measurements for the relevant dates. Cannot contain NA-values.}

\item{bandwidth}{A \code{double}. The bandwidth used in the kernel smoothing}
}
\value{
A \code{double}, which is the cross-validation value for the fit given the \code{bandwidth}.
}
\description{
Local average kernel smoothing loo cv of a set of measure values
}
\details{
Local Average Kernel Smoothing Leave One Out Cross Validation

Measurement values with weight < 0.01 is not included for efficiency
}
\examples{
\dontrun{

  print(1)

}
}
