% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llks}
\alias{llks}
\title{Local Linear Kernel Smoothing Method Using A Gaussian Kernel Function}
\arguments{
\item{date}{A \code{numeric} vector containing dates for the smoothing.}

\item{median}{A \code{numeric} vector contaning measurements for the relevant dates. Can contain NA-values, which will be removed if present.}

\item{bandwidth}{A \code{double}. The bandwidth used in the kernel smoothing}

\item{average}{A \code{double} representing the sample mean of relevant measurements. If set to below \code{0}, it will be calculated in this function instead.}

\item{standard_deviation}{A \code{double} representing the sample standard deviation of relevant measurements. If set to below \code{0}, it will be calculated in this function instead.}

\item{approximate}{A \code{logical} value. If set to \code{TRUE}, gradients are estimated via second differences. If set to \code{FALSE}, a plug-in estimator of the gradients is used.}

\item{matrix_approach}{A \code{loigcal} value. Should matrix calculus be used to smooth. Only relevant if \code{method} is \code{ll}. Matrix calculus is faster, but may not be stable in some cases.}
}
\value{
A \code{list} containing kernel smoothed values.
}
\description{
Local linear kernel smoothing of a set of measure values
}
\details{
Local Linear Kernel Smoothing Method Using A Gaussian Kernel Function

Measurement values with weight < 0.01 is not included for efficiency
}
\examples{
\dontrun{

  print(1)

}
}
