% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moving_median}
\alias{moving_median}
\title{Moving Median Smoothing Method}
\arguments{
\item{date}{A \code{numeric} vector containing dates for the smoothing.}

\item{median}{A \code{numeric} vector contaning measurements for the relevant dates. Can contain NA-values, which will be removed if present.}

\item{bandwidth}{A \code{double}. The bandwidth used in the kernel smoothing}
}
\value{
A \code{list} containing moving median smoothed values.
}
\description{
Moving median smoothing of a set of measure values
}
\details{
Moving Median Smoothing Method

No details.
}
\examples{
\dontrun{

  print(1)

}
}
