% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_median_smoothing.R
\name{moving_median_smoothing}
\alias{moving_median_smoothing}
\title{Moving Median Curve for Multiple Instruments Within Multiple Laboratories}
\usage{
moving_median_smoothing(
  data,
  by,
  measure = c("median", "hyper", "hypo"),
  bw = 30,
  attach = TRUE
)
}
\arguments{
\item{data}{A \code{data.table} containing measurements.}

\item{by}{A \code{character} vector containing the names of the grouping variables.}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{bw}{An \code{integer}. The window width used for the moving median smoothing.}

\item{attach}{A \code{logical} value. Should we attach the information found in \code{data} to the output?}
}
\value{
A \code{data.table} containing the smoothed values for multiple instruments within multiple laboratories
}
\description{
Moving Median Curve for Multiple Instruments Within Multiple Laboratories
}
\examples{
print(1)
}
