% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_median_smoothing.R
\name{moving_median_smoothing0}
\alias{moving_median_smoothing0}
\title{Moving Median Curve for One Instrument Within One Laboratory}
\usage{
moving_median_smoothing0(
  data,
  measure = c("median", "hyper", "hypo"),
  bw = 30,
  attach = TRUE
)
}
\arguments{
\item{data}{A \code{data.table} containing measurements.}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{bw}{An \code{integer}. The window width used for the moving median smoothing.}

\item{attach}{A \code{logical} value. Should we attach the information found in \code{data} to the output?}
}
\value{
A \code{data.table} containing the smoothed values for one instrument within one laboratory
}
\description{
Moving Median Curve for One Instrument Within One Laboratory
}
\examples{
print(1)
}
