% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduce_bias_warnings}
\alias{reduce_bias_warnings}
\title{Reduce Bias Type Warnings}
\arguments{
\item{warnings}{A \code{logical} vector containing raw bias type warnings.}

\item{snooze}{A \code{integer}. How many days should be used in snoozing.}
}
\value{
A \code{logical} vector containing a reduced set of warnings after snoozing.
}
\description{
Reduce bias type warnings using the snoozing functionality
}
\details{
Reduce Bias Type Warnings

No details.
}
\examples{
\dontrun{

  print(1)

}
}
