% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduce_slope_warnings}
\alias{reduce_slope_warnings}
\title{Reduce Slope Type Warnings}
\arguments{
\item{data}{A \code{list} containing slope warning data.}

\item{stringent}{A \code{logical} value. Should slope warnings be reduced?}
}
\value{
A \code{list} with updated warnings
}
\description{
Reduce slope type warnings using the slope functionality
}
\details{
Reduce Slope Type Warnings

No details.
}
\examples{
\dontrun{

  print(1)

}
}
