% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_warning_process_p.R
\name{slope_warning}
\alias{slope_warning}
\title{Slope warning process}
\usage{
slope_warning(
  data,
  from,
  end_dates,
  method,
  measure,
  approximate = TRUE,
  tol = 2,
  dur = 3,
  stringent_slope_warning = TRUE,
  bw = 11,
  attach = TRUE
)
}
\arguments{
\item{data}{A \code{data.table} containing measurements.}

\item{from}{A \code{IDate} value. The start of the warning process.}

\item{end_dates}{A \code{IDate} vector. The endings of the warning process.}

\item{method}{A \code{character} signifying which smoothing method that is used. Valid options include \code{lc} (Local-Average kernel smoothing) and \code{ll} (Local-Linear kernel smoothing)}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{approximate}{A \code{logical} value. If set to \code{TRUE}, second differences are used to estimate gradients. If \code{FALSE} other methods are utilized.}

\item{tol}{A \code{double} value. What is the warning threshold for the smoothed curve?}

\item{dur}{A \code{integer} value. How many days must the gradient be above the threshold given in \code{tol} in order to trigger a warning}

\item{stringent_slope_warning}{A \code{logical} value. If set to \code{FALSE}, all triggered warnings are displayed. If set to \code{TRUE}, we avoid excessive number of warnings. See \code{?reduce_slope_warnings()} for more information.}

\item{bw}{A \code{double}. The bandwidth used for the smoothing.}

\item{attach}{A \code{logical} value. Should we attach the information found in \code{data} to the output?}
}
\value{
A \code{data.table} object containing the slope warning process information
}
\description{
Slope warning process
}
\examples{
print(1)
}
