% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_process_plot.R
\name{warning_process_plot}
\alias{warning_process_plot}
\title{Plot a Warning Process}
\usage{
warning_process_plot(
  data,
  from,
  to,
  measure = c("median", "hyper", "hypo"),
  warning = c("slope", "bias", "peer_group"),
  strip_percentage = 1,
  plot = FALSE,
  plotting_arguments = list(title = "default", subtitle = "default", xlab = "default",
    ylab = "default", legend_show = "default", legend_position = "default",
    date_tick_distance = "default", unit = "default", raw_data_shape = "default",
    raw_data_alpha = "default", raw_data_size = "default", smooth_data_linewidth =
    "default", warning_segment_color = "default", warning_segment_linewidth = "default",
    warning_point_shape = "default", warning_point_color = "default", warning_point_size
    = "default")
)
}
\arguments{
\item{data}{A \code{data.table} containing necessary warning data.}

\item{from}{A \code{IDate} object that is the first date where a warning is calculated.}

\item{to}{A \code{IDate} object that is the last date where a warning is calculated}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{warning}{A \code{character} signifying which warning process that is of interest to simulate.}

\item{strip_percentage}{A \code{double} between 0 and 100. Data outside are stripped from the plot.}

\item{plot}{A \code{logical} value. Set to \code{TRUE} to plot the plot right away. Set to \code{FALSE} to store the plot.}

\item{plotting_arguments}{A \code{list} of optional plotting arguments.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a Warning Process
}
\examples{
print(1)
}
