% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_warning_process_x.R
\name{xgenerate_warnings}
\alias{xgenerate_warnings}
\title{Alternative Generation of All Warnings for All Instruments Within Each Laboratory}
\usage{
xgenerate_warnings(
  data,
  from = "2022-01-02",
  to = "2023-01-01",
  method = c("lc", "ll"),
  tol_median_slope = 2,
  tol_hyper_slope = 8,
  tol_hypo_slope = 8,
  tol_median_bias = 10,
  tol_hyper_bias = 10,
  tol_hypo_bias = 10,
  tol_pg = 10,
  dur = 3,
  bw1 = 11,
  bw2 = 30,
  bw3 = 365,
  only_warnings = FALSE
)
}
\arguments{
\item{data}{A \code{data.table} containing measurements.}

\item{from}{A \code{IDate} value. The start of the warning process, i.e., the first date where a warning is calculated.}

\item{to}{A \code{IDate} value. The end of the warning process, i.e., the last date where a warning is calculated.}

\item{method}{A \code{character} signifying which kernel smoothing method that is used for the slope warnings. Valid options include \code{lc} (Local-Average kernel smoothing) and \code{ll} (Local-Linear kernel smoothing).}

\item{tol_median_slope}{A \code{double} value. What is the warning threshold for the median slope warning?}

\item{tol_hyper_slope}{A \code{double} value. What is the warning threshold for the hyper slope warning?}

\item{tol_hypo_slope}{A \code{double} value. What is the warning threshold for the hypo slope warning?}

\item{tol_median_bias}{A \code{double} value. What is the warning threshold for the median month-year bias warning?}

\item{tol_hyper_bias}{A \code{double} value. What is the warning threshold for the hyper month-year bias warning?}

\item{tol_hypo_bias}{A \code{double} value. What is the warning threshold for the hypo month-year bias warning?}

\item{tol_pg}{A \code{double} value. What is the warning threshold for the median peer group bias warning?}

\item{dur}{An \code{integer} value. How many days must the gradient be above the threshold given in \code{tol} in order to trigger a slope warning. Will be the same for all measures for the slope warning. The default is \code{3}.}

\item{bw1}{A \code{double} value. The bandwidth used for the kernel smoothing. Typically set to 11.}

\item{bw2}{An \code{integer} value. The bandwidth used for monthly moving median and the monthly median for peer group data. Typically set to 30.}

\item{bw3}{An \code{integer} value. The bandwidth used for yearly moving median. Typically set to 365.}

\item{only_warnings}{A \code{logical} value. If set to \code{TRUE}, only the dates with warnings generated are outputted. If set to \code{FALSE}, the whole process it outputted.}
}
\value{
A \code{data.table} object containing the warning process information for each of the seven relevant warnings.
The output columns for each type of warning may be slightly different, so the output \code{data.table} from this function will contain a lot for columns.
}
\description{
Alternative Generation of All Warnings for All Instruments Within Each Laboratory
}
\examples{
print(1)
}
