% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_warning_process_x.R
\name{xwarning_process0}
\alias{xwarning_process0}
\title{Alternative Warning Process for One Instrument Within One Laboratory}
\usage{
xwarning_process0(
  data,
  pg_data = NULL,
  from = "2021-01-01",
  to = "2022-12-31",
  measure = c("median", "hyper", "hypo"),
  method = c("lc", "ll"),
  warning = c("slope", "bias", "peer_group"),
  tol = 4,
  dur = 3,
  bw1 = 11,
  bw2 = 30,
  bw3 = 365,
  snooze = 0L,
  stringent_slope_warning = FALSE,
  attach = TRUE,
  approximate = TRUE
)
}
\arguments{
\item{data}{A \code{data.table} containing measurements.}

\item{pg_data}{A \code{data.table} containing peer group measurements.}

\item{from}{A \code{IDate} value. The date before the start of the warning process.}

\item{to}{A \code{IDate} value. The date before the end of the warning process.}

\item{measure}{A \code{character} signifying which measure to smooth.}

\item{method}{A \code{character} signifying which smoothing method that is used. Valid options include \code{lc} (Local-Average kernel smoothing) and \code{ll} (Local-Linear kernel smoothing). Only relevant if \code{warning} is set to \code{slope}.}

\item{warning}{A \code{character} signifying which warning process that is of interest to simulate.}

\item{tol}{A \code{double} value. What is the warning threshold for relevant \code{warning}?}

\item{dur}{A \code{integer} value. How many days must the gradient be above the threshold given in \code{tol} in order to trigger a warning. Only relevant if \code{warning} is set to \code{slope}.}

\item{bw1}{A \code{double}. The bandwidth used for the kernel smoothing. Typically set to 11.}

\item{bw2}{A \code{double}. The bandwidth used for monthly moving median. Typically set to 30.}

\item{bw3}{A \code{double}. The bandwidth used for yearly moving median. Typically set to 365.}

\item{snooze}{A \code{integer} signifying the number of days recurring warnings are snoozed. Only relevant if \code{warning} is \code{bias} or \code{peer_group}. Set to \code{0} to disable snoozing.}

\item{stringent_slope_warning}{A \code{logical} value. If set to \code{FALSE}, all triggered warnings are displayed. If set to \code{TRUE}, we avoid excessive number of warnings. See \code{?reduce_slope_warnings()} for more information. Only relevant if \code{warning} is set to \code{slope}.}

\item{attach}{A \code{logical} value. Should we attach the information found in \code{data} to the output?}

\item{approximate}{A \code{logical} value. If set to \code{TRUE}, second differences are used to estimate gradients. If \code{FALSE} other methods are utilized. Only relevant if \code{warning} is set to \code{slope}.}
}
\value{
A \code{data.table} object containing the warning process information
}
\description{
Alternative Warning Process for One Instrument Within One Laboratory
}
\examples{
print(1)
}
